+++ beem/rc.py
@@ -6,7 +6,7 @@ from __future__ import unicode_literals
 import logging
 import json
 from .instance import shared_steem_instance
-from beem.constants import state_object_size_info
+from beem.constants import state_object_size_info, resource_execution_time, EXEC_FOLLOW_CUSTOM_OP_SCALE
 import hashlib
 from binascii import hexlify, unhexlify
 import os
@@ -42,13 +42,14 @@ class RC(object):
         tx_size = len(txWire)
         return tx_size
 
-    def get_resource_count(self, tx_size, state_bytes_count=0, new_account_op_count=0, market_op_count=0):
+    def get_resource_count(self, tx_size, execution_time_count, state_bytes_count=0, new_account_op_count=0, market_op_count=0):
         """Creates the resource_count dictionary based on tx_size, state_bytes_count, new_account_op_count and market_op_count"""
         resource_count = {"resource_history_bytes": tx_size}
         resource_count["resource_state_bytes"] = state_object_size_info["transaction_object_base_size"]
         resource_count["resource_state_bytes"] += state_object_size_info["transaction_object_byte_size"] * tx_size
         resource_count["resource_state_bytes"] += state_bytes_count
         resource_count["resource_new_accounts"] = new_account_op_count
+        resource_count["resource_execution_time"] = execution_time_count
         if market_op_count > 0:
             resource_count["resource_market_bytes"] = tx_size
         return resource_count
@@ -83,7 +84,8 @@ class RC(object):
         state_bytes_count = state_object_size_info["comment_object_base_size"]
         state_bytes_count += state_object_size_info["comment_object_permlink_char_size"] * permlink_length
         state_bytes_count += state_object_size_info["comment_object_parent_permlink_char_size"] * parent_permlink_length
-        resource_count = self.get_resource_count(tx_size, state_bytes_count)
+        execution_time_count = resource_execution_time["comment_operation_exec_time"]
+        resource_count = self.get_resource_count(tx_size, execution_time_count, state_bytes_count)
         return self.steem.get_rc_cost(resource_count)
 
     def vote_dict(self, vote_dict):
@@ -110,7 +112,8 @@ class RC(object):
     def vote(self, tx_size=210):
         """Calc RC for a vote"""
         state_bytes_count = state_object_size_info["comment_vote_object_base_size"]
-        resource_count = self.get_resource_count(tx_size, state_bytes_count)
+        execution_time_count = resource_execution_time["vote_operation_exec_time"]
+        resource_count = self.get_resource_count(tx_size, execution_time_count, state_bytes_count)
         return self.steem.get_rc_cost(resource_count)
 
     def transfer_dict(self, transfer_dict):
@@ -139,7 +142,8 @@ class RC(object):
 
     def transfer(self, tx_size=290, market_op_count=1):
         """Calc RC of a transfer"""
-        resource_count = self.get_resource_count(tx_size, market_op_count=market_op_count)
+        execution_time_count = resource_execution_time["transfer_operation_exec_time"]
+        resource_count = self.get_resource_count(tx_size, execution_time_count, market_op_count=market_op_count)
         return self.steem.get_rc_cost(resource_count)
 
     def custom_json_dict(self, custom_json_dict):
@@ -168,20 +172,57 @@ class RC(object):
         """
         op = operations.Custom_json(**custom_json_dict)
         tx_size = self.get_tx_size(op)
-        return self.custom_json(tx_size=tx_size)
+        follow_id = custom_json_dict["id"] == "follow"
+        return self.custom_json(tx_size=tx_size, follow_id=follow_id)
 
-    def custom_json(self, tx_size=444):
-        resource_count = self.get_resource_count(tx_size)
+    def custom_json(self, tx_size=444, follow_id=False):
+        execution_time_count = resource_execution_time["custom_json_operation_exec_time"]
+        if follow_id:
+            execution_time_count *= EXEC_FOLLOW_CUSTOM_OP_SCALE
+        resource_count = self.get_resource_count(tx_size, execution_time_count)
         return self.steem.get_rc_cost(resource_count)
 
     def account_update_dict(self, account_update_dict):
         """Calc RC costs for account update"""
         op = operations.Account_update(**account_update_dict)
         tx_size = self.get_tx_size(op)
-        resource_count = self.get_resource_count(tx_size)
+        execution_time_count = resource_execution_time["account_update_operation_exec_time"]
+        resource_count = self.get_resource_count(tx_size, execution_time_count)
         return self.steem.get_rc_cost(resource_count)
 
     def claim_account(self, tx_size=300):
         """Claim account"""
-        resource_count = self.get_resource_count(tx_size, new_account_op_count=1)
+        execution_time_count = resource_execution_time["claim_account_operation_exec_time"]
+        resource_count = self.get_resource_count(tx_size, execution_time_count, new_account_op_count=1)
+        return self.steem.get_rc_cost(resource_count)
+
+    def get_authority_byte_count(self, auth):
+        return (state_object_size_info["authority_base_size"]
+                + state_object_size_info["authority_account_member_size"] * len(auth["account_auths"])
+                + state_object_size_info["authority_key_member_size"] * len(auth["key_auths"]))
+
+    def account_create_dict(self, account_create_dict):
+        """Calc RC costs for account create"""
+        op = operations.Account_create(**account_create_dict)
+        state_bytes_count = state_object_size_info["account_object_base_size"]
+        state_bytes_count += state_object_size_info["account_authority_object_base_size"]
+        state_bytes_count += self.get_authority_byte_count(account_create_dict["owner"])
+        state_bytes_count += self.get_authority_byte_count(account_create_dict["active"])
+        state_bytes_count += self.get_authority_byte_count(account_create_dict["posting"])
+        tx_size = self.get_tx_size(op)
+        execution_time_count = resource_execution_time["account_update_operation_exec_time"]
+        resource_count = self.get_resource_count(tx_size, execution_time_count, state_bytes_count)
+        return self.steem.get_rc_cost(resource_count)
+
+    def create_claimed_account_dict(self, create_claimed_account_dict):
+        """Calc RC costs for claimed account create"""
+        op = operations.Create_claimed_account(**create_claimed_account_dict)
+        state_bytes_count = state_object_size_info["account_object_base_size"]
+        state_bytes_count += state_object_size_info["account_authority_object_base_size"]
+        state_bytes_count += self.get_authority_byte_count(create_claimed_account_dict["owner"])
+        state_bytes_count += self.get_authority_byte_count(create_claimed_account_dict["active"])
+        state_bytes_count += self.get_authority_byte_count(create_claimed_account_dict["posting"])
+        tx_size = self.get_tx_size(op)
+        execution_time_count = resource_execution_time["account_update_operation_exec_time"]
+        resource_count = self.get_resource_count(tx_size, execution_time_count, state_bytes_count)
         return self.steem.get_rc_cost(resource_count)